correcao de exercicios

+++ exx030.py
@@ -1,4 +1,4 @@
-velo = int(input('Qual a velocidade atual do carro? '))
+velo = float(input('Qual a velocidade atual do carro? '))
 if velo > 80:
     print('MULTADO! Você excedeu o limite permitido que é de 80Km/h')
     print('Você deve pagar uma multa de R$ {:.2f}'.format((velo - 80) * 7))
